Improve test login UI and UX; Bump version to 1.0.30
The first time I used the test login, I didn't even see the username and
PIN code input fields, I only saw a "Continue" button, which I pressed.
The app didn't complain about it incorrect credentials, it just
"continued", but the test login wasn't active.

App Store review also apparently couldn't figure out the UI (they
couldn't find "Test IdP" during login, so the app was still using the
production backend), and the app was rejected in review.

I improved the test login UI (and UX):
- added a header ("E-kirjasto test login"), so that you have some idea
  where you landed from the deep link
- added a background color to the input fields, so that they don't
  disappear into the app's background
- added labels to the input fields, because the hint text alone isn't
  enough in my opinion, and the hint text has horrible contrast anyway
- changed "Continue" to "Login" to make it more clear what will happen
- trying to login with incorrect credentials now shows an error, instead
  of quietly "resetting" back to the production backend
- added an explicit "Cancel" button, if you want to exit the test login
  (and switch back to using the production backend)

There was also another issue with the test login, where the test login
page showed up for less than a second, before the regular onboarding
or login page would open on top of it, so that you were unable to
actually use the test login. I wasn't able to find a reliable way to
reproduce it, but it happened often enough that it might have also been
the reason for the app review rejection (all that the reviewer really
said was that they couldn't find the "Test IdP" option dugin login, so
*something* went wrong during test login, but it's unclear exactly what
that was).

Yet another issue was that after the test login, the app didn't always
go to the login page, but went straight to the main UI instead, possibly
without even changing the backend. Trying to login through the settings
crashed the app in this case, and app reviewers aren't being instructed
to login through there anyway. This should be fixed now, so the regular
login page (and onboarding page) should always show up after the test
login now.

I also changed the test login credentials, so that they're not
literally "test" and "1234", in case someone happens to stumble upon
the deep link. I still hate that they're just hard-coded, especially in
a *public* Git repo (hello to any random people reading this). But this
PR is big enough as it is, so let's fix that a bit later (but still
soon, I'll work on the build automation and secrets soon anyway).

I also added a "docs" directory and added instructions for App Store
reviewers there (to be copied into App Store Connect). The new
instructions reflect the changes made in this PR.

Unrelated to the test login, I also switched the order of the tabs in
the app to match the order on Android. The old order was a bit weird in
my opinion, because the first tab was related to books, then there was
a tab for magazines, then 2 more tabs related to books, and then the
settings tab. With the new order, the first 3 tabs are related to
books, and then there are tabs for magazines and settings. This way
you're not "skipping over" the magazines tab when you're handling books
(browsing, loaning, reserving, etc.).