Add CI workflows, Fastlane uploads, build configs
This adds CI workflows, Fastlane signing and uploads, and build
configurations for each backend (production, beta, dev, and ellibs).

This is also a complete rewrite of all of the build scripts, since the
old ones were quite specific to Palace, and E-kirjasto has changed some
things.

I also rewrote all of the documentation, since it was also about Palace,
and not updated for E-kirjasto.

The documentation has a pretty good description of everything, so I
won't rewrite everything in this commit/PR text.

Issue: SIMPLYE-381